chore: merging uniswap versions
