+++ src/lib/Uniswap/contracts.ts
@@ -1,6 +1,10 @@
 import { ethers } from 'ethers';
 import { CHAIN_ID, LISTEN_FOR_TRANSACTIONS } from '../../enums';
 import { IProtocolContractDefinitions } from '../../types';
+import PoolAbi from './abis/Pool.json';
+import RouterAbi from './abis/Router.json';
+import SwapRouter01Abi from './abis/SwapRouter01.json';
+import SwapRouter02Abi from './abis/SwapRouter02.json';
 import UniswapUniversalRouterABI from './abis/UniswapUniversalRouter.json';
 import UniswapV2RouterABI from './abis/UniswapV2Router.json';
 
@@ -9,7 +13,52 @@ export enum CONTRACT_ENUM {
   ROUTER_V2 = 'RouterV2',
   POOL_V3 = 'PoolV3',
   ROUTER_V3 = 'RouterV3',
-  UNIVERSAL_ROUTER = 'UniversalRouter'
+  UNIVERSAL_ROUTER = 'UniversalRouter',
+}
+
+export enum COMMAND_ENUM {
+  // V3 Swap Commands
+  V3_SWAP_EXACT_IN = 0x00,
+  V3_SWAP_EXACT_OUT = 0x01,
+  
+  // V2 Swap Commands  
+  V2_SWAP_EXACT_IN = 0x02,
+  V2_SWAP_EXACT_OUT = 0x03,
+  
+  // Permit2 Commands
+  PERMIT2_PERMIT = 0x0a,
+  PERMIT2_TRANSFER_FROM = 0x0b,
+  
+  // Wrap/Unwrap Commands
+  WRAP_ETH = 0x0c,
+  UNWRAP_WETH = 0x0d,
+  
+  // Sweep Commands
+  SWEEP = 0x0e,
+  TRANSFER = 0x0f,
+  
+  // Payment Commands  
+  PAY_PORTION = 0x10,
+  
+  // V3 LP Commands
+  V3_MINT = 0x11,
+  V3_COLLECT = 0x12,
+  V3_BURN = 0x13,
+  
+  // V2 LP Commands
+  V2_MINT = 0x14,
+  V2_BURN = 0x15,
+  
+  // Routing Commands
+  ROUTE = 0x16,
+  NOOP = 0x17,
+  
+  // Timestamp Commands
+  TIMESTAMP = 0x18,
+  
+  // Limit Order Commands
+  LIMIT_ORDER = 0x19,
+  CANCEL_LIMIT_ORDER = 0x1a
 }
 
 export enum EVENT_ENUM {
@@ -67,6 +116,53 @@ export const contracts: IProtocolContractDefinitions = {
       [CHAIN_ID.BASE]: {
         address: '0x3fC91A3afd70395Cd496C647d5a6CC9D4B2b7FAD',
         listenForTransactions: [LISTEN_FOR_TRANSACTIONS.INCOMING]
+      }
+    },
+    events: {}
+  },
+  [CONTRACT_ENUM.POOL_V3]: {
+    interface: new ethers.Interface(PoolAbi.abi),
+    deployments: {
+      [CHAIN_ID.ARBITRUM]: {
+        address: ethers.ZeroAddress,
+        listenForTransactions: [LISTEN_FOR_TRANSACTIONS.INCOMING]
+      }
+    },
+    events: {
+      [EVENT_ENUM.SWAP]: {
+        signature: '0xc42079f94a6350d7e6235f29174924f928cc2ac818eb64fed8004e115fbcca67',
+        abi: new ethers.Interface([
+          'event Swap(address indexed sender, address indexed recipient, int256 amount0, int256 amount1, uint160 sqrtPriceX96, uint128 liquidity, int24 tick)'
+        ])
+      }
+    }
+  },
+  [CONTRACT_ENUM.ROUTER_V3]: {
+    interface: new ethers.Interface(RouterAbi.abi),
+    deployments: {
+      [CHAIN_ID.ARBITRUM]: {
+        address: '0x68b3465833fb72A70ecDF485E0e4C7bD8665Fc45',
+        listenForTransactions: [LISTEN_FOR_TRANSACTIONS.OUTGOING]
+      }
+    },
+    events: {}
+  },
+  'SwapRouter01': {
+    interface: new ethers.Interface(SwapRouter01Abi.abi),
+    deployments: {
+      [CHAIN_ID.ARBITRUM]: {
+        address: '0xE592427A0AEce92De3Edee1F18E0157C05861564',
+        listenForTransactions: [LISTEN_FOR_TRANSACTIONS.OUTGOING]
+      }
+    },
+    events: {}
+  },
+  'SwapRouter02': {
+    interface: new ethers.Interface(SwapRouter02Abi.abi),
+    deployments: {
+      [CHAIN_ID.ARBITRUM]: {
+        address: '0x68b3465833fb72A70ecDF485E0e4C7bD8665Fc45',
+        listenForTransactions: [LISTEN_FOR_TRANSACTIONS.OUTGOING]
       }
     },
     events: {}
